Move installation into plugmngr
